re-named & re-organized files/folders to be more consistant
