+++ src/components/bar_chart/BarChart.js
