--- src/components/footer.css
@@ -1,7 +1,0 @@
-.footer {
-  height: 5vh;
-  margin: 1.5%;
-}
-.copy {
-  margin-left: 2%;
-}
