+++ src/components/sign-up.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import './form.css';
+import '../styles/form.css';
 import { SignUpFormDiv } from '../styles/FormStyle';
 import { axiosWithAuth } from '../utils/axiosWithAuth';
 const initialVal = {
