+++ src/index.js
@@ -1,33 +1,26 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import { createStore } from 'redux';
-import { Provider } from 'react-redux';
-import { rootReducer } from './state/reducers/index.js';
-
 import { ReactQueryDevtools } from 'react-query-devtools';
+
 import './index.css';
 import 'antd/dist/antd.less';
 
-import { Container } from './components/Container';
+import NavBar from '../src/components/NavBar';
 import About from './components/about';
 import RegisterForm from './components/sign-up';
 import LoginForm from './components/sign-in';
-import Dashboard from './components/Dashboard';
-import NavBar from '../src/components/NavBar';
+import PieChartView from './components/pie_chart/PieChartView';
+import BarChartView from './components/bar_chart/BarChartView';
 import Footer from './components/Footer';
-import DashboardChart from './components/DashboardChart';
-
-const store = createStore(rootReducer);
+import MapView from './components/map/MapView.js';
 
 ReactDOM.render(
-  <Provider store={store}>
-    <Router>
-      <React.StrictMode>
-        <App />
-      </React.StrictMode>
-    </Router>
-  </Provider>,
+  <Router>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  </Router>,
   document.getElementById('root')
 );
 
@@ -35,7 +28,6 @@ function App() {
   return (
     <>
       <div>
-        {/* Add the Route in to App */}
         <header>
           <NavBar />
         </header>
@@ -47,17 +39,17 @@ function App() {
           <Route path="/login">
             <LoginForm />
           </Route>
-          <Route path="/dashboard">
-            <Dashboard />
-          </Route>
-          <Route path="/dashboard-chart">
-            <DashboardChart />
-          </Route>
           <Route path="/about">
             <About />
           </Route>
+          <Route path="/charts-by-force">
+            <PieChartView />
+          </Route>
+          <Route path="/charts-by-state">
+            <BarChartView />
+          </Route>
           <Route path="/">
-            <Container />
+            <MapView />
           </Route>
         </Switch>
         <Footer />
